SE-27: Rename file to reflect correct functionality.
